refactor: replace count and average timeseries builders with unified ReportBuilder
