+++ app/builders/v2/reports/timeseries/base_timeseries_builder.rb
@@ -1,12 +1,13 @@
 class V2::Reports::Timeseries::BaseTimeseriesBuilder
   include TimezoneHelper
   include DateRangeHelper
+
   DEFAULT_GROUP_BY = 'day'.freeze
 
   pattr_initialize :account, :params
 
   def scope
-    case params[:type].to_sym
+    case dimension_type.to_sym
     when :account
       account
     when :inbox
@@ -18,6 +19,21 @@ class V2::Reports::Timeseries::BaseTimeseriesBuilder
     when :team
       team
     end
+  end
+
+  def data_source
+    @data_source ||= Reports::DataSource.for(
+      account: account,
+      metric: params[:metric],
+      dimension_type: dimension_type,
+      dimension_id: params[:id],
+      scope: scope,
+      range: range,
+      group_by: group_by,
+      timezone: timezone,
+      timezone_offset: params[:timezone_offset],
+      business_hours: params[:business_hours]
+    )
   end
 
   def inbox
@@ -41,6 +57,12 @@ class V2::Reports::Timeseries::BaseTimeseriesBuilder
   end
 
   def timezone
-    @timezone ||= timezone_name_from_offset(params[:timezone_offset])
+    @timezone ||= timezone_name_from_params(params[:timezone], params[:timezone_offset])
+  end
+
+  private
+
+  def dimension_type
+    (params[:type].presence || 'account').to_s
   end
 end
